requirements.txt
python-dotenv==1.0.1
PyPDF2==3.0.1
python-docx==1.1.0
google-genai
nltk==3.8.1
spacy==3.7.4
